initial typescript converstion
