+++ .eslintrc.js
@@ -1,4 +1,8 @@
 module.exports = {
   root: true,
   extends: '@react-native-community',
-};
+  extends: 'standard',
+  rules: {
+    semi: [1, 'never'],
+  },
+}
