+++ assets/images.js
@@ -1,20 +1,20 @@
 const images = {
-    arrow: require('./img/arrow.png'),
-    arrowDown: require('./img/arrow-down.png'),
-    infoBlue: require('./img/info-blue.png'),
-    infoGray: require('./img/info-gray.png'),
-    infoWhite: require('./img/info-white.png'),
-    logo: require('./img/logo.png'),
-    logoLarge: require('./img/logo-large.png'),
-    logoText: require('./img/logo-text.png'),
-    navConnect: require('./img/nav-connect.png'),
-    navContacts: require('./img/nav-contacts.png'),
-    navCredentials: require('./img/nav-credentials.png'),
-    navHome: require('./img/nav-home.png'),
-    navSettings: require('./img/nav-settings.png'),
-    receive: require('./img/receive.png'),
-    send: require('./img/send.png'),
-    waiting: require('./img/waiting.png'),
-  }
-  
-  export default images
+  arrow: require('./img/arrow.png'),
+  arrowDown: require('./img/arrow-down.png'),
+  infoBlue: require('./img/info-blue.png'),
+  infoGray: require('./img/info-gray.png'),
+  infoWhite: require('./img/info-white.png'),
+  logo: require('./img/logo.png'),
+  logoLarge: require('./img/logo-large.png'),
+  logoText: require('./img/logo-text.png'),
+  navConnect: require('./img/nav-connect.png'),
+  navContacts: require('./img/nav-contacts.png'),
+  navCredentials: require('./img/nav-credentials.png'),
+  navHome: require('./img/nav-home.png'),
+  navSettings: require('./img/nav-settings.png'),
+  receive: require('./img/receive.png'),
+  send: require('./img/send.png'),
+  waiting: require('./img/waiting.png'),
+}
+
+export default images
