+++ components/AppHeader/index.tsx
@@ -16,7 +16,11 @@ import Images from '../../assets/images'
 import AppStyles from '../../assets/styles'
 import Styles from './styles'
 
-function AppHeader(props) {
+interface AppHeaderProps {
+  headerText?: string
+}
+
+function AppHeader (props: AppHeaderProps) {
   let history = useHistory()
 
   return (
@@ -28,16 +32,16 @@ function AppHeader(props) {
         <View style={Styles.headerView}>
           {props.headerText ? (
             <>
-            <Image source={Images.logo} style={{marginRight: 10}} />
-            <Text
-              style={[
-                AppStyles.h1,
-                AppStyles.textSecondary,
-                Styles.textBorder,
-                AppStyles.textUpper,
-              ]}>
-              {props.headerText}
-            </Text>
+              <Image source={Images.logo} style={{marginRight: 10}} />
+              <Text
+                style={[
+                  AppStyles.h1,
+                  AppStyles.textSecondary,
+                  Styles.textBorder,
+                  AppStyles.textUpper,
+                ]}>
+                {props.headerText}
+              </Text>
             </>
           ) : (
             <Image source={Images.logoText} style={{}} />
