+++ components/PinEnter/index.tsx
@@ -18,8 +18,8 @@ import {useHistory} from 'react-router-native'
 
 import * as Keychain from 'react-native-keychain'
 
-import AppHeaderLarge from '../AppHeaderLarge/index.js'
-import LoadingOverlay from '../LoadingOverlay/index.js'
+import AppHeaderLarge from '../AppHeaderLarge/index'
+import LoadingOverlay from '../LoadingOverlay/index'
 
 import AppStyles from '../../assets/styles'
 
@@ -99,7 +99,11 @@ function PinEnter(props) {
             />
           </View>
           <TouchableOpacity
-            style={[AppStyles.button, AppStyles.backgroundPrimary, {marginTop: 30}]}
+            style={[
+              AppStyles.button,
+              AppStyles.backgroundPrimary,
+              {marginTop: 30},
+            ]}
             onPress={() => {
               Keyboard.dismiss()
               checkPin(pin)
