+++ package.json
@@ -33,6 +33,11 @@
     "@babel/runtime": "^7.12.5",
     "@commitlint/cli": "^11.0.0",
     "@react-native-community/eslint-config": "^2.0.0",
+    "@types/jest": "^26.0.20",
+    "@types/react": "^16.14.4",
+    "@types/react-native": "^0.63.50",
+    "@types/react-router-native": "^5.1.0",
+    "@types/react-test-renderer": "^16.9.5",
     "babel-jest": "^26.6.3",
     "commitlint": "^11.0.0",
     "eslint": "^7.12.1",
@@ -40,7 +45,8 @@
     "jest": "^26.6.3",
     "metro-react-native-babel-preset": "^0.63.0",
     "prettier": "^2.1.2",
-    "react-test-renderer": "16.9.0"
+    "react-test-renderer": "16.9.0",
+    "typescript": "^4.2.2"
   },
   "husky": {
     "hooks": {
