+++ types.ts
@@ -1,0 +1,15 @@
+export interface IContact {
+    label: string
+    sublabel: string
+    address: string
+    phone: string
+    email: string
+}
+
+export interface ICredential {
+    label: string
+    sublabel: string
+    first_name: string
+    last_name: string
+    credential_date: string
+}
